+++ tests/extract_wiki_format_test.py
@@ -21,7 +21,7 @@ class TestWikiFormatExtracts(unittest.TestCase):
 
     def test_title_after_fetching(self):
         page = self.wiki.article('Test_1')
-        page.structured()
+        page._fetch_structured()
         self.assertEqual(page.title, 'Test 1')
 
     def test_summary(self):